test: reduce code duplication after splitting test files in cloud/aws-sdk/v3/s3
